stock-ai-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
aiomysql==0.2.0
pymysql==1.1.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1

# 주식 데이터
pykrx==1.0.47
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
finance-datareader>=0.9.50

# AI — 멀티 LLM
anthropic==0.26.0
openai==1.30.0
google-generativeai==0.7.0

# 비동기 작업
celery==5.4.0
redis==5.0.4
celery[redis]

# 유틸
python-dotenv==1.0.1
httpx==0.27.0
beautifulsoup4==4.12.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.3
